网上书城基本代码实现

+++ src/main/java/com/zym/bookstore/web/CriteriaBook.java
@@ -3,7 +3,7 @@ package com.zym.bookstore.web;
 public class CriteriaBook {
 
 	private float minPrice = 0;
-	private float maxPrice = Integer.MAX_VALUE;
+	private float maxPrice = Float.MAX_VALUE;
 
 	private int pageNo;
